Adding alterations to script, stored procedures,entities, repository and services

+++ src/main/java/com/acmeairlines/airlinesim/repository/AircraftRepository.java
@@ -5,7 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AircraftRepository extends JpaRepository<Aircraft, Integer> {
-
-    // Define any custom query methods here
+public interface AircraftRepository extends JpaRepository<Aircraft, Long> {
 }
