+++ src/main/java/com/acmeairlines/airlinesim/repository/AirlineRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AirlineRepository extends JpaRepository<Airline, Integer> {
-    // Define any custom query methods here
+public interface AirlineRepository extends JpaRepository<Airline, Long> {
 }
