+++ src/main/java/com/acmeairlines/airlinesim/repository/AirportRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AirportRepository extends JpaRepository<Airport, Integer> {
-    // Define any custom query methods here
+public interface AirportRepository extends JpaRepository<Airport, Long> {
 }
