+++ src/main/java/com/acmeairlines/airlinesim/repository/SeatAssignmentRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeatAssignmentRepository extends JpaRepository<SeatAssignment, Integer> {
-    // Define any custom query methods here
+public interface SeatAssignmentRepository extends JpaRepository<SeatAssignment, Long> {
 }
